+++ zpi_aspnet_test/zpi_aspnet_test.Tests/Matchers/DoesNotThrow.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-namespace zpi_aspnet_test.Tests
+namespace zpi_aspnet_test.Tests.Matchers
 {
-	public class DoesNotThrow
+	public static class DoesNotThrow
 	{
 		public static DoesNotThrowMatcher<T> An<T>() where T : Exception
 		{
